--- src/components/bs5/header/_search.functions.js
@@ -1,2 +1,0 @@
-// Must be wrapped in form class site-search
-// Included in header
