+++ src/components/bs5/header/header.hbs
@@ -879,10 +879,12 @@
             </div>
             {{/if}}
             {{#if siteSearchAsset.value}}
-            <div class="qld__header__site-search qld__header__site-search--closed">
+            <div id="qld-header-search" class="qld__header__site-search qld__header__site-search--closed">
                 <form class="site-search" role="search" 
                     {{#if siteSearchAsset.formAction.url}}
                         action="{{siteSearchAsset.formAction.url}}"
+                    {{else}}
+                        action="https://www.qld.gov.au/search"
                     {{/if}}>
                         {{{ searchInput }}}
                 </form>
