+++ src/components/bs5/searchInput/search.json
