+++ src/main.js
@@ -3,7 +3,9 @@ import { videoEmbedPlay, videoTranscriptTitle } from "./components/bs5/video/vid
 import { initializeNavbar } from './components/bs5/navbar/navbar.functions';
 import { breadcrumbShorten } from "./components/bs5/breadcrumbs/breadcrumb.functions";
 import { positionQuickExit, initQuickexit } from './components/bs5/quickexit/quickexit.functions';
-import { toggleSearch, showSuggestions } from './components/bs5/header/header.functions';
+
+import { toggleSearch } from './components/bs5/header/header.functions';
+import { showSuggestions } from './components/bs5/searchInput/search.functions';
 
 window.addEventListener('scroll', positionQuickExit, true);
 window.addEventListener('resize', positionQuickExit, true);
@@ -15,10 +17,14 @@ window.addEventListener("DOMContentLoaded", () => {
 
     //Header search
     let headerSearchButton = document.querySelector('.qld__main-nav__toggle-search');
-    if(headerSearchButton) {
+
+    if (headerSearchButton) {
       document.querySelector('.qld__main-nav__toggle-search').addEventListener('click', toggleSearch);
     }
-    const searchInput = document.getElementById('search-input');
+
+    const form = document.querySelector('.site-search');
+    const searchInput = form.querySelector('.qld-search-input input');
+
     if (searchInput) {
       let timeout;
 
@@ -36,12 +42,6 @@ window.addEventListener("DOMContentLoaded", () => {
       searchInput.addEventListener('click', function() {
         if (this.value === '') {
           showSuggestions('', true);
-        }
-      });
-      // Close suggestions when clicking outside
-      document.addEventListener('click', function(event) {
-        if (!searchInput.contains(event.target) && !document.getElementById('suggestions').contains(event.target)) {
-          document.getElementById('suggestions').style.display = 'none';
         }
       });
     }
